+++ src/ARISDefaults.m
@@ -52,7 +52,7 @@
   //Just load immutable settings from root.plist to find default defaults, and put them in easily accessible dictionary
   NSDictionary *settingsDict  = [NSDictionary dictionaryWithContentsOfFile:[[[[NSBundle mainBundle] bundlePath] stringByAppendingPathComponent:@"Settings.bundle"] stringByAppendingPathComponent:@"Root.plist"]];
   NSArray *prefDictArray = settingsDict[@"PreferenceSpecifiers"];
-  for(int i = 0; i < prefDictArray.count; i++)
+  for(long i = 0; i < prefDictArray.count; i++)
   {
     if(prefDictArray[i][@"DefaultValue"]) 
         defaultDefaults[prefDictArray[i][@"Key"]] = prefDictArray[i][@"DefaultValue"];
@@ -87,7 +87,7 @@
   [defaults setBool:NO forKey:@"clearCache"];
   [defaults synchronize];
 
-  showPlayerOnMap        = [defaults boolForKey:@"showPlayerOnMap"];
+  showPlayerOnMap = [defaults boolForKey:@"showPlayerOnMap"];
 
   User *u          = [[User alloc] init];
   u.user_id        = [defaults integerForKey:@"user_id"];
