+++ src/game_pickers/ARISStarView.h
@@ -9,7 +9,7 @@
 #import <UIKit/UIKit.h>
 
 @interface ARISStarView : UIView
-- (void) setSpacing:(int)s;
-- (void) setRating:(int)r;
-- (int) rating;
+- (void) setSpacing:(long)s;
+- (void) setRating:(long)r;
+- (long) rating;
 @end
