+++ src/models/tags/TagsModel.h
@@ -15,9 +15,9 @@
 }
 
 - (NSArray *) tags;
-- (Tag *) tagForId:(int)tag_id;
-- (ObjectTag *) objectTagForId:(int)object_tag_id;
-- (NSArray *) tagsForObjectType:(NSString*)t id:(int)object_id;
+- (Tag *) tagForId:(long)tag_id;
+- (ObjectTag *) objectTagForId:(long)object_tag_id;
+- (NSArray *) tagsForObjectType:(NSString*)t id:(long)object_id;
 - (void) requestTags;
 - (void) clearGameData;
